+++ packages/react-pacer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @tanstack/react-pacer
 
+## 0.22.0
+
+### Minor Changes
+
+- feat: upgrade to tanstack store with useSelector hook ([`c9f6385`](https://github.com/TanStack/pacer/commit/c9f63850f816bfc95eca74699b0fff45bca02867))
+
+### Patch Changes
+
+- Updated dependencies [[`c9f6385`](https://github.com/TanStack/pacer/commit/c9f63850f816bfc95eca74699b0fff45bca02867)]:
+  - @tanstack/pacer@0.21.0
+
 ## 0.21.1
 
 ### Patch Changes